fix dependency bugs
