+++ paddle/cinn/ir/ir_analyzer/ir_analyzer.cc
@@ -27,6 +27,7 @@
 #include "paddle/cinn/ir/ir_printer.h"
 #include "paddle/cinn/ir/ir_visitor.h"
 #include "paddle/cinn/ir/schedule/ir_schedule.h"
+#include "paddle/cinn/ir/schedule/ir_schedule_util.h"
 #include "paddle/cinn/ir/schedule/schedule_base.h"
 #include "paddle/cinn/ir/schedule/schedule_desc.h"
 #include "paddle/cinn/ir/tensor.h"
@@ -37,144 +38,6 @@
 namespace cinn {
 namespace ir {
 namespace analyzer {
-namespace {
-
-struct FindBlocksVisitor {
-  explicit FindBlocksVisitor(const std::string& block_name = "")
-      : block_name_(block_name) {}
-
-  std::vector<Expr> operator()(const Expr* expr) {
-    Visit(expr);
-    return result;
-  }
-
- private:
-  void Visit(const Expr* expr) {
-    if (!expr->defined()) return;
-    if (!block_name_.empty() && !result.empty()) return;
-    if (expr->As<ir::For>()) {
-      Visit(&(expr->As<ir::For>()->body));
-    } else if (expr->As<ir::ScheduleBlockRealize>()) {
-      if (!expr->As<ir::ScheduleBlockRealize>()->iter_values.empty()) {
-        auto* schedule_block = expr->As<ir::ScheduleBlockRealize>()
-                                   ->schedule_block.As<ir::ScheduleBlock>();
-        if (block_name_.empty() || schedule_block->name == block_name_) {
-          result.emplace_back(*expr);
-        }
-      } else {
-        Visit(&(expr->As<ir::ScheduleBlockRealize>()->schedule_block));
-      }
-    } else if (expr->As<ir::ScheduleBlock>()) {
-      Visit(&(expr->As<ir::ScheduleBlock>()->body));
-    } else if (expr->As<ir::Block>()) {
-      for (auto& n : expr->As<ir::Block>()->stmts) Visit(&n);
-    } else if (expr->As<ir::IfThenElse>()) {
-      Visit(&(expr->As<ir::IfThenElse>()->true_case));
-      Visit(&(expr->As<ir::IfThenElse>()->false_case));
-    }
-  }
-  std::string block_name_;
-  std::vector<Expr> result{};
-};
-
-struct FindLoopsVisitor {
-  explicit FindLoopsVisitor(const Expr& block) : block_(block) {}
-
-  std::vector<Expr> operator()(const Expr* expr) {
-    CHECK(block_.As<ir::ScheduleBlockRealize>());
-    visit_end = false;
-    Visit(expr);
-    return result;
-  }
-
- private:
-  void Visit(const Expr* expr) {
-    if (visit_end || !expr->defined()) return;
-    if (expr->As<ir::For>()) {
-      father_loops.emplace_back(*expr);
-      Visit(&(expr->As<ir::For>()->body));
-      father_loops.pop_back();
-    } else if (expr->As<ir::ScheduleBlockRealize>()) {
-      if (!expr->As<ir::ScheduleBlockRealize>()->iter_values.empty() &&
-          (*expr == block_)) {
-        result = father_loops;
-        visit_end = true;
-        return;
-      } else {
-        Visit(&(expr->As<ir::ScheduleBlockRealize>()->schedule_block));
-      }
-    } else if (expr->As<ir::ScheduleBlock>()) {
-      Visit(&(expr->As<ir::ScheduleBlock>()->body));
-    } else if (expr->As<ir::Block>()) {
-      for (auto& n : expr->As<ir::Block>()->stmts) Visit(&n);
-    } else if (expr->As<ir::IfThenElse>()) {
-      Visit(&(expr->As<ir::IfThenElse>()->true_case));
-      Visit(&(expr->As<ir::IfThenElse>()->false_case));
-    }
-  }
-
-  std::vector<Expr> father_loops{};
-  std::vector<Expr> result{};
-  bool visit_end{false};
-  const Expr& block_;
-};
-
-struct FindBlockParent : public ir::IRMutator<> {
- public:
-  explicit FindBlockParent(const std::string& block_name)
-      : block_name_(block_name) {}
-
-  void operator()(Expr* expr) { IRMutator::Visit(expr, expr); }
-
- private:
-  void Visit(const ir::Block* expr, Expr* op) override {
-    if (target_) return;
-    for (auto& stmt : expr->stmts) {
-      if (stmt.As<ir::ScheduleBlockRealize>()) {
-        if (stmt.As<ir::ScheduleBlockRealize>()
-                ->schedule_block.As<ir::ScheduleBlock>()
-                ->name == block_name_) {
-          target_ = op;
-          return;
-        }
-      }
-    }
-    IRMutator::Visit(expr, op);
-  }
-
-  void Visit(const ir::For* expr, Expr* op) override {
-    if (target_) return;
-    if (expr->body.As<ir::ScheduleBlockRealize>()) {
-      if (expr->body.As<ir::ScheduleBlockRealize>()
-              ->schedule_block.As<ir::ScheduleBlock>()
-              ->name == block_name_) {
-        target_ = op;
-        return;
-      }
-    }
-    IRMutator::Visit(expr, op);
-  }
-
-  void Visit(const ir::ScheduleBlock* expr, Expr* op) override {
-    if (target_) return;
-    if (expr->body.As<ir::ScheduleBlockRealize>()) {
-      if (expr->body.As<ir::ScheduleBlockRealize>()
-              ->schedule_block.As<ir::ScheduleBlock>()
-              ->name == block_name_) {
-        target_ = op;
-        return;
-      }
-    }
-    IRMutator::Visit(expr, op);
-  }
-
-  std::string block_name_;
-
- public:
-  ir::Expr* target_{nullptr};
-};
-
-}  // namespace
 
 bool HasBlock(const std::vector<Expr>& exprs, const std::string& block_name) {
   for (auto& it_expr : exprs) {
