--- paddle/cinn/ir/schedule/utils/CMakeLists.txt
@@ -1,3 +1,0 @@
-core_gather_headers()
-
-gather_srcs(cinnapi_src SRCS ir_schedule_util.cc)
