--- paddle/cinn/ir/schedule/utils/ir_schedule_util.cc
@@ -1,28 +1,0 @@
-// Copyright (c) 2022 CINN Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#pragma once
-
-#include "paddle/cinn/ir/schedule/utils/ir_schedule_util.h"
-
-#include <map>
-#include <random>
-#include <string>
-#include <unordered_map>
-#include <utility>
-#include <vector>
-
-namespace cinn {
-namespace ir {}  // namespace ir
-}  // namespace cinn
